Refactor the autoconfiguration mechanims. I've studied the source code of the project https://github.com/spring-projects/spring-boot/blob/master/spring-boot-autoconfigure to understand how spring boot organize it's classes.
Therefore, the new class structure was made based components like
@ConfigurationProperties, @Conditional and subclasses of SpringBootCondition,
like AnyNestedCondition and AllNestedConditions.

By default, if no feedback is provided, the metrics analyzed by thresno will be
automatically configured to use the console for feedback. Otherwise,
it'll be disabled. If you want sysout the analysis set the value of this property to true.